Use XDG_DATA_HOME and security fix
Summary:
The EncFS has security issues when the encrypted files are shared
in the open. For instance on a usb-pendrive or a shared drive.

Only when the user picks EncFS we then continue to not allow the user to pick his 'device' directory where the encrypted files would go, just store this on the XDG_DATA_HOME which is defined as;
 the base directory relative to which user specific data files should be stored

Users can continue picking their datadir just fine when they pick the CryFS and other future backends.

BUG:385982

Reviewers: ivan, #plasma

Reviewed By: ivan, #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D8469