Have a default backend (if one available)
Summary:
The creation of a new vault uses a wizard.

The first page of the wizard currently shows a combobox with 2 backends, with one semi-randomly selected by default.
The user is instructed to pick one.

Following the bugreport on one of the backend packages being scary to install on neon, I'm proposing we make the wizard a little more wizardly. But without losing any functionality. Users that want can still do all they can do now.

We make it easier for most users while not limiting the advanced ones.

BUG: 385971

Reviewers: ivan, #plasma

Subscribers: nicolasfella, ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D8449