[applet] Restore former default action
The default was changed in the ExpandibleListItem port to "Open with
file manager" in the interests of having better defaults, but this
change was ill-considered for the following reasons:
1. It left no way to only open the vault without also opening a file
   manager, which apparently is a real and valid use case
2. The "Open with file manager" item in the expanded view became totally
   redundant

I still think that opening in the file manager is a better default, but
that change needs to be done properly, and doing so requires string
and icon changes and further adjustments in the applet which are not
suitable for 5.19. So for now, let's revert to the pre-5.19 behavior to
unbreak people.

BUG: 422917
FIXED-IN: 5.19.1