[Applet] Simplify actions and improve defaults
Right now the default action is to unlock the vault but not show it in
the file manager, which is a valid use case but probably not the best
default. Most of the time when you open your vault, you want to somehow
directly interact with its contents. Accordingly, "Unlock and Open" is
made the default action, and "Unlock" (without showing" is made a
non-default action in the expanded actions view.

Additionally, the "Lock Forcefully" and "Unlock" actions are made
mutually exclusive such that only one is showing, because the other is
always irrelevant, depending on the state of the vault.