[applet] Add "Show in File Manager" action when opened
After recent changes to the actions, when you open a vault, there is no
standalone action to show it in a file manager. Normally this is not
needed if you use the "Unlock and Open" action, but if you use that and
then close it by mistake, or use the "Unlock" action that does not
automatically spawn a file manager window, it may not be obvious how to
actually get to the vault.

Accordingly, this commit adds a standalone "Show in File Manager" action
that appears in the expanded view, but only when the vault is currently
unlocked.