Put "Add New Vault" button in header for System Tray representation
Over the past few Plasma releases, we've moved towards generally putting
applets' actions into the header in the form of icons-only toolbuttons.
A requirement is that the icon be fairly universally recognizable, which
is the case for the "plus sign" icon used to add a new vault.
Accordingly, let's move the button into the header for the system tray
representation. The button remains in its current location for panel or
desktop representations.

Internally, this is done by defining a new action--which is
automatically consumed by the system tray and displayed in the
header--and making the other "New vault" buttons pull their text and
actions from it.