Remove .directory file (if exists) before mount
Dolphin has a tendency to create a .directory file in directories
on a users system without them really being aware of that.
Plasma vault doesn't like this and would refuse to mount there.

Check the contents of the dir (should be empty) and if the only item
there is the dot-directroy, delete it.