Remove applet list item animations
This was always the indended appearance, but it was being overridden by
a bug in the ExpandableListItem itself. Let's remove the animations here
so that the bug in question can be fixed while preserving the intended
visual effect.

(cherry picked from commit 233e004b9564aea24719c4e7129fc5d6fc981ca7)