Make applet active in the System Tray only when any vaults are open
Right now the applet is active when any vaults *exist*, and inactive
otherwise. This lets you always access a vault with two clicks.

As a downside, the Vaults icon is always taking up space in the visible
part of your System Tray. My sense is that most users of this feature
aren't opening and closing vaults all day every day, so for these
people, showing the icon 100% of the time wastes space.

Accordingly, let's make the applet active only when any vaults are
actually open. This makes the visible contents of the System Tray more
contextually relevant, and also provides a visual reminder when you've
forgotten to close your last vault--the icon is still visible!

The drawback to this change is that accessing a vault when none are
already open becomes a three-click task. I think this is an acceptable
trade-off for the above benefits, and people who prefer the status quo
can restore it by setting the applet to be always visible in their System
Tray settings.

Implementation-wise, the change relies on storing a list of open vaults
in the model that's kept in sync with the one in the KDED service, which
mirrors what's already done for the list of busy vaults and vaults in an
error state.

CCBUG: 476423