kded: only open vault in new Dolphin window if needed
The code deliberately forces a new Dolphin window to prevent issues
when the opened vault is activity-specific and there's already a
Dolphin window open in a different activity.

However none of this is necessary if there's only one actvity or the
vault being opened is not activity-specific. Handle those cases too so
we can use the normal behavior of opening a new tab more often.

BUG: 502239
FIXED-IN: 6.4.0